[FEATURE] Recursive update of page - moved methods
