+++ Classes/IndexQueue/RecordMonitor.php
@@ -27,8 +27,6 @@ namespace ApacheSolrForTypo3\Solr\IndexQueue;
 use ApacheSolrForTypo3\Solr\AbstractDataHandlerListener;
 use ApacheSolrForTypo3\Solr\Domain\Index\Queue\RecordMonitor\Helper\MountPagesUpdater;
 use ApacheSolrForTypo3\Solr\GarbageCollector;
-use ApacheSolrForTypo3\Solr\IndexQueue\Queue;
-use ApacheSolrForTypo3\Solr\System\Configuration\TypoScriptConfiguration;
 use ApacheSolrForTypo3\Solr\System\TCA\TCAService;
 use ApacheSolrForTypo3\Solr\Util;
 use TYPO3\CMS\Backend\Utility\BackendUtility;
@@ -43,13 +41,6 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
  */
 class RecordMonitor extends AbstractDataHandlerListener
 {
-
-    /**
-     * Solr TypoScript configuration
-     *
-     * @var TypoScriptConfiguration
-     */
-    protected $solrConfiguration;
 
     /**
      * Index Queue
@@ -412,32 +403,6 @@ class RecordMonitor extends AbstractDataHandlerListener
     }
 
     /**
-     * Retrieves a record, taking into account the additionalWhereClauses of the
-     * Indexing Queue configurations.
-     *
-     * @param string $recordTable Table to read from
-     * @param int $recordUid Id of the record
-     * @return array Record if found, otherwise empty array
-     */
-    protected function getRecord($recordTable, $recordUid)
-    {
-        $record = [];
-        $indexingConfigurations = $this->solrConfiguration->getEnabledIndexQueueConfigurationNames();
-
-        foreach ($indexingConfigurations as $indexingConfigurationName) {
-            $record = $this->getRecordWhenIndexConfigurationIsValid($recordTable, $recordUid, $indexingConfigurationName);
-            if (!empty($record)) {
-                // if we found a record which matches the conditions, we can continue
-                break;
-            }
-        }
-
-        return $record;
-    }
-
-    // Handle pages showing content from another page
-
-    /**
      * Triggers Index Queue updates for other pages showing content from the
      * page currently being updated.
      *
@@ -477,96 +442,4 @@ class RecordMonitor extends AbstractDataHandlerListener
         return $pid;
     }
 
-    /**
-     * Retrieves the name of the  Indexing Queue Configuration for a record
-     *
-     * @param string $recordTable Table to read from
-     * @param int $recordUid Id of the record
-     * @return string Name of indexing configuration
-     */
-    protected function getIndexingConfigurationName($recordTable, $recordUid)
-    {
-        $name = $recordTable;
-        $indexingConfigurations = $this->solrConfiguration->getEnabledIndexQueueConfigurationNames();
-        foreach ($indexingConfigurations as $indexingConfigurationName) {
-            if (!$this->solrConfiguration->getIndexQueueConfigurationIsEnabled($indexingConfigurationName)) {
-                // ignore disabled indexing configurations
-                continue;
-            }
-
-            $record = $this->getRecordWhenIndexConfigurationIsValid($recordTable, $recordUid, $indexingConfigurationName);
-            if (!empty($record)) {
-                $name = $indexingConfigurationName;
-                // FIXME currently returns after the first configuration match
-                break;
-            }
-        }
-
-        return $name;
-    }
-
-    /**
-     * This method return the record array if the table is valid for this indexingConfiguration.
-     * Otherwise an empty array will be returned.
-     *
-     * @param string $recordTable
-     * @param integer $recordUid
-     * @param string $indexingConfigurationName
-     * @return array
-     */
-    protected function getRecordWhenIndexConfigurationIsValid($recordTable, $recordUid, $indexingConfigurationName)
-    {
-        if (!$this->isValidTableForIndexConfigurationName($recordTable, $indexingConfigurationName)) {
-            return [];
-        }
-
-        $recordWhereClause = $this->solrConfiguration->getIndexQueueAdditionalWhereClauseByConfigurationName($indexingConfigurationName);
-
-        if ($recordTable === 'pages_language_overlay') {
-            return $this->getPageOverlayRecordWhenParentIsAccessible($recordUid, $recordWhereClause);
-        }
-
-        return (array)BackendUtility::getRecord($recordTable, $recordUid, '*', $recordWhereClause);
-    }
-
-    /**
-     * This method retrieves the pages_language_overlay record when the parent record is accessible
-     * through the recordWhereClause
-     *
-     * @param int $recordUid
-     * @param string $parentWhereClause
-     * @return array
-     */
-    protected function getPageOverlayRecordWhenParentIsAccessible($recordUid, $parentWhereClause)
-    {
-        $overlayRecord = (array)BackendUtility::getRecord('pages_language_overlay', $recordUid, '*');
-        $pageRecord = (array)BackendUtility::getRecord('pages', $overlayRecord['pid'], '*', $parentWhereClause);
-
-        if (empty($pageRecord)) {
-            return [];
-        }
-
-        return $overlayRecord;
-    }
-
-    /**
-     * This method is used to check if a table is an allowed table for an index configuration.
-     *
-     * @param string $recordTable
-     * @param string $indexingConfigurationName
-     * @return boolean
-     */
-    protected function isValidTableForIndexConfigurationName($recordTable, $indexingConfigurationName)
-    {
-        $tableToIndex = $this->solrConfiguration->getIndexQueueTableNameOrFallbackToConfigurationName($indexingConfigurationName);
-
-        $isMatchingTable = $tableToIndex === $recordTable;
-        $isPagesPassedAndOverlayRequested = $tableToIndex === 'pages' && $recordTable === 'pages_language_overlay';
-
-        if ($isMatchingTable || $isPagesPassedAndOverlayRequested) {
-            return true;
-        }
-
-        return false;
-    }
 }
